Replace settings modal with settings page
