+++ resources/less/admin.less
@@ -1,3 +1,3 @@
-.FoFDefaultGroupSettingsModal .Modal-body {
-  overflow-y: auto;
+.FoFDefaultGroupSettingsPage {
+  margin-top: 25px;
 }
